\name{IButils-package}
\alias{IButils-package}
\alias{IButils}
\docType{package}
\title{
\packageTitle{IButils}
}
\description{
\packageDescription{IButils}
}
\details{
  The package contains functions to download historical data from IB.
}
\author{
\packageAuthor{IButils}

Maintainer: \packageMaintainer{IButils}
}
\keyword{ package }
