\name{latest_timestamp}
\alias{latest_timestamp}
\title{
  Latest Timestamp
}
\description{
  Get the latest timestamp from the files in a directory.
}
\usage{
latest_timestamp(directory, id)
}
\arguments{
  \item{directory}{
    character
}
  \item{id}{
    character
}
}
\details{

  Get the latest (i.e. largest) timestamp for
  particular \code{id}.

}
\value{
  numeric
}
\author{
  Enrico Schumann
}
\seealso{
  \code{\link{ib_hist_data}}
}
\examples{
## TODO
}
