\name{IButils-package}
\alias{IButils-package}
\alias{IButils}
\docType{package}
\title{
\packageTitle{IButils}
}
\description{
\packageDescription{IButils}
}
\details{

  Functionality for the \acronym{TWS} \acronym{API}; for
  instance, for downloading historical data or retrieving
  current positions.


}
\author{
\packageAuthor{IButils}

Maintainer: \packageMaintainer{IButils}
}
\keyword{package}
